common/pyproject.toml
[project]
name = "axiom"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=45.0.4",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.0",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.41",
]
